requirements.txt
ffmpeg-python~=0.2
numpy~=1.23
Pillow~=9.3
